+++ CHANGELOG.md
@@ -26,12 +26,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed typo in RoomEncryptionEnabledViewModel
 - Upgrade Gradle to 8.9
 - Fixed Java toolchain configuration in emoji subproject
+- Only create media directory in export when needed
+- Export view model exposes more info in error message
+- Export uses better timestamp formats
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Export is not working, when size of exported events did not match a multiple of buffer size
+- Fix wrong filename in export
 
 ### Security